added alert dialogs for update and delete buttons in account settings

+++ newstation-front-end/src/pages/AccountSettings/index.tsx
@@ -8,6 +8,7 @@ import useStyles from '../../styles/_Dashboard';
 import { logout, getUser, getCompanyId, getEmployeeId } from '../../utils';
 import { State, Company } from './types';
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 const AccountSettings = () => {
   const classes = useStyles();
@@ -74,23 +75,60 @@ const AccountSettings = () => {
   }
 
   const getAccountEmployee = async () => {
-    const id = getEmployeeId();
-    const response = await axios.get("/employee/" + id);
-    const { data } = response;
-    setEditedAccount(data)
+    try {
+      const id = getEmployeeId();
+      const response = await axios.get("/employee/" + id);
+      const { data } = response;
+      setEditedAccount(data)
+    }
+    catch (error) {
+      return { "message": "Error Retrieving Employee!" };
+    }
   }
 
   const getAccountCompany = async () => {
-    const id = getCompanyId();
-    const response = await axios.get("/company/" + id);
-    const { result } = response.data
-    setEditedCompany(result)
+    try {
+      const id = getCompanyId();
+      const response = await axios.get("/company/" + id);
+      const { result } = response.data
+      setEditedCompany(result)
+    }
+    catch (error) {
+      return { "message": "Error Retrieving Company!" };
+    }
   }
 
   const handleUpdateAccount = async () => {
     try {
-      const { id } = editedAccount
-      const result = await axios.put('/employee/' + id, editedAccount)
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "This Employee will be Updated!.",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: 'Yes, update it!',
+        cancelButtonText: 'No, cancel!',
+        reverseButtons: true
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          const { id } = editedAccount
+          const result = await axios.put('/employee/' + id, editedAccount)
+          Swal.fire(
+            'Deleted!',
+            'Account Updated!',
+            'success'
+          )
+        } else if (
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          Swal.fire(
+            'Cancelled',
+            'Cancelled Account Update',
+            'error'
+          )
+        }
+      })
     } catch (error) {
       return { "message": "Error Update!" };
     }
@@ -98,8 +136,36 @@ const AccountSettings = () => {
 
   const handleUpdateCompany = async () => {
     try {
-      const { id } = editedAccount
-      const result = await axios.put('/company/' + id, editedAccount)
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "This Company will be Updated!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: 'Yes, update it!',
+        cancelButtonText: 'No, cancel!',
+        reverseButtons: true
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          const { id } = editedAccount
+          const result = await axios.put('/company/' + id, editedAccount)
+          Swal.fire(
+            'Deleted!',
+            'Account Updated!',
+            'success'
+          )
+        } else if (
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          Swal.fire(
+            'Cancelled',
+            'Cancelled Account Update',
+            'error'
+          )
+        }
+      })
+
     } catch (error) {
       return { "message": "Error Update!" };
     }
@@ -107,11 +173,39 @@ const AccountSettings = () => {
 
   const handleDeleteAccount = async () => {
     try {
-      const { id } = editedAccount
-      const result = await axios.delete('/employee/' + id);
-      logout();
-      history.push('/login');
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "Your account will be deleted.",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'No, cancel!',
+        reverseButtons: true
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          const { id } = editedAccount
+          const result = await axios.delete('/employee/' + id);
 
+          Swal.fire(
+            'Deleted!',
+            'Account Deleted!',
+            'success'
+          )
+
+          logout();
+          history.push('/login');
+        } else if (
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          Swal.fire(
+            'Cancelled',
+            'Cancelled Account Deletion',
+            'error'
+          )
+        }
+      })
     } catch (error) {
       return { "message": "Error Delete!" };
     }
@@ -119,11 +213,38 @@ const AccountSettings = () => {
 
   const handleDeleteCompany = async () => {
     try {
-      const { id } = editedCompany
-      const result = await axios.delete('/company/' + id);
-      logout();
-      history.push('/login');
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "This Company will be deleted.",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'No, cancel!',
+        reverseButtons: true
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          const { id } = editedCompany
+          const result = await axios.delete('/company/' + id);
 
+          Swal.fire(
+            'Deleted!',
+            'Account Deleted!',
+            'success'
+          )
+          logout();
+          history.push('/login');
+        } else if (
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          Swal.fire(
+            'Cancelled',
+            'Cancelled Account Deletion',
+            'error'
+          )
+        }
+      })
     } catch (error) {
       return { "message": "Error Delete!" };
     }
